refactor(core): Merge parse r4 and r4 mappings files
